libraries
ssl libraries for windows
changed the output directory
moved windows and mac libraries


Former-commit-id: 44e36756e427501dd3632ffb9db27096badd6fc6 [formerly 3880d9b1801f40f2a81f37322011713e2be6e910]
Former-commit-id: ccbe05bbf74e0aa4e09063a1e8efa47e927222d9
Former-commit-id: 2ee084a22bf1f2c5f14c56ac271922082e8e143e [formerly 1591892530c1edeb52e7b522b009486e5156046f]
Former-commit-id: 0e2472715fbeba57ae63278c6f61c014083a4032

+++ sources/clavier/pianoscene.h
@@ -96,7 +96,7 @@ private:
     void triggerNoteOn(const int note, const int vel);
     void triggerNoteOff(const int note);
     static qreal sceneWidth(int startKey, int keys);
-    static float getOffsetX(int numKey);
+    static double getOffsetX(int numKey);
     static QString formatSequence(QKeySequence sequence);
 
     int m_numKeys;
